combine fast_bootstrap code into bootstrap.py
